VAULT-6727 use the correct file
