+++ path_login_test.go
@@ -12,6 +12,7 @@ import (
 	"encoding/pem"
 	"errors"
 	"fmt"
+	"github.com/hashicorp/vault/sdk/helper/tokenutil"
 	"net/http"
 	"reflect"
 	"strings"
@@ -1472,5 +1473,60 @@ func requireErrorCode(t *testing.T, err error, expectedCode int) {
 		t.Fatal("err was not logical.HTTPCodedError")
 	default:
 		t.Fatalf("wrong error code, expected %d, got %d", expectedCode, codedErr.Code())
+	}
+}
+
+func TestResolveRole(t *testing.T) {
+	b, storage := getBackend(t)
+	role := "testrole"
+
+	validRoleStorageEntry := &roleStorageEntry{
+		TokenParams: tokenutil.TokenParams{
+			TokenPolicies:   []string{"test"},
+			TokenPeriod:     3 * time.Second,
+			TokenTTL:        1 * time.Second,
+			TokenMaxTTL:     5 * time.Second,
+			TokenNumUses:    12,
+			TokenBoundCIDRs: nil,
+		},
+		Policies:                 []string{"test"},
+		Period:                   3 * time.Second,
+		ServiceAccountNames:      []string{"name"},
+		ServiceAccountNamespaces: []string{"namespace"},
+		TTL:                      1 * time.Second,
+		MaxTTL:                   5 * time.Second,
+		NumUses:                  12,
+		BoundCIDRs:               nil,
+		AliasNameSource:          aliasNameSourceDefault,
+	}
+
+	entry, err := logical.StorageEntryJSON("role/"+role, validRoleStorageEntry)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := storage.Put(context.TODO(), entry); err != nil {
+		t.Fatal(err)
+	}
+
+	loginData := map[string]interface{}{
+		"role": role,
+	}
+	loginReq := &logical.Request{
+		Operation: logical.ResolveRoleOperation,
+		Path:      "login",
+		Storage:   storage,
+		Data:      loginData,
+		Connection: &logical.Connection{
+			RemoteAddr: "127.0.0.1",
+		},
+	}
+
+	resp, err := b.HandleRequest(context.Background(), loginReq)
+	if err != nil || (resp != nil && resp.IsError()) {
+		t.Fatalf("err:%v resp:%#v", err, resp)
+	}
+
+	if resp.Data["role"] != role {
+		t.Fatalf("Role was not as expected. Expected %s, received %s", role, resp.Data["role"])
 	}
 }
